# Conflicts:
#	src/main/java/edu/tust/neusoft/backend/model/dto/GoodsDTO.java

+++ src/main/java/edu/tust/neusoft/backend/service/CartsService.java
@@ -4,8 +4,8 @@ import edu.tust.neusoft.backend.model.Carts;
 import edu.tust.neusoft.backend.response.Result;
 
 public interface CartsService {
-    Result getCarts(int userId);
+    Result getCarts(Long userId);
     Result addCart(Carts cart);
     Result updateCart(Carts cart);
-    Result deleteCart(Long cartId);
+    Result deleteCart(Integer cartId);
 }
